Refactor auth helpers into dpc-testing module
Moved the auth code out of dpc-api and into dpc-testing
