+++ dpc-api/src/test/java/gov/cms/dpc/api/auth/jwt/JWTUnitTests.java
@@ -14,6 +14,7 @@ import gov.cms.dpc.macaroons.MacaroonBakery;
 import gov.cms.dpc.macaroons.config.TokenPolicy;
 import gov.cms.dpc.macaroons.store.MemoryRootKeyStore;
 import gov.cms.dpc.macaroons.thirdparty.MemoryThirdPartyKeyStore;
+import gov.cms.dpc.testing.APIAuthHelpers;
 import gov.cms.dpc.testing.BufferedLoggerHandler;
 import io.dropwizard.testing.junit5.DropwizardExtensionsSupport;
 import io.dropwizard.testing.junit5.ResourceExtension;
@@ -56,7 +57,7 @@ class JWTUnitTests {
 
     private JWTUnitTests() {
         try {
-            JWTKeys.put("correct", APITestHelpers.generateKeyPair());
+            JWTKeys.put("correct", APIAuthHelpers.generateKeyPair());
         } catch (NoSuchAlgorithmException e) {
             e.printStackTrace();
         }
@@ -207,7 +208,7 @@ class JWTUnitTests {
         @Test
         void testMissingJWTPublicKey() throws NoSuchAlgorithmException {
             // Submit JWT with missing key
-            final KeyPair keyPair = APITestHelpers.generateKeyPair();
+            final KeyPair keyPair = APIAuthHelpers.generateKeyPair();
 
             final String jwt = Jwts.builder()
                     .setHeaderParam("kid", "wrong")
@@ -261,7 +262,7 @@ class JWTUnitTests {
 
         @Test
         void testJWTWrongSigningKey() throws NoSuchAlgorithmException {
-            final KeyPair keyPair = APITestHelpers.generateKeyPair();
+            final KeyPair keyPair = APIAuthHelpers.generateKeyPair();
 
             final String jwt = Jwts.builder()
                     .setHeaderParam("kid", "correct")
