+++ dpc-api/src/test/java/gov/cms/dpc/api/validations/ProfileTests.java
@@ -9,6 +9,7 @@ import ca.uhn.fhir.rest.server.exceptions.UnprocessableEntityException;
 import gov.cms.dpc.api.APITestHelpers;
 import gov.cms.dpc.api.AbstractSecureApplicationTest;
 import gov.cms.dpc.fhir.DPCIdentifierSystem;
+import gov.cms.dpc.testing.APIAuthHelpers;
 import org.hl7.fhir.dstu3.model.*;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -28,7 +29,7 @@ class ProfileTests extends AbstractSecureApplicationTest {
 
     @Test
     void testPatientProfile() throws IOException, URISyntaxException {
-        final IGenericClient client = APITestHelpers.buildAuthenticatedClient(ctx, getBaseURL(), ORGANIZATION_TOKEN, KEY_ID, privateKey);
+        final IGenericClient client = APIAuthHelpers.buildAuthenticatedClient(ctx, getBaseURL(), ORGANIZATION_TOKEN, KEY_ID, privateKey);
         // Create a new patient record
 
         final Patient invalidPatient = new Patient();
@@ -90,7 +91,7 @@ class ProfileTests extends AbstractSecureApplicationTest {
 
     @Test
     void testProviderProfile() throws IOException, URISyntaxException {
-        final IGenericClient client = APITestHelpers.buildAuthenticatedClient(ctx, getBaseURL(), ORGANIZATION_TOKEN, KEY_ID, privateKey);
+        final IGenericClient client = APIAuthHelpers.buildAuthenticatedClient(ctx, getBaseURL(), ORGANIZATION_TOKEN, KEY_ID, privateKey);
 
         final Practitioner invalidPractitioner = new Practitioner();
         invalidPractitioner.addName().addGiven("Test").setFamily("Practitioner");
@@ -147,7 +148,7 @@ class ProfileTests extends AbstractSecureApplicationTest {
     @Disabled
         // Disabled until DPC-614 and DPC-616 are merged.
     void testAttributionProfile() throws IOException, URISyntaxException {
-        final IGenericClient client = APITestHelpers.buildAuthenticatedClient(ctx, getBaseURL(), ORGANIZATION_TOKEN, KEY_ID, privateKey);
+        final IGenericClient client = APIAuthHelpers.buildAuthenticatedClient(ctx, getBaseURL(), ORGANIZATION_TOKEN, KEY_ID, privateKey);
 
         final Group invalidGroup = new Group();
         invalidGroup.addMember().setEntity(new Reference("Patient/strange-patient"));
